cast type

+++ openrlhf/models/actor.py
@@ -91,7 +91,7 @@ class Actor(nn.Module):
         #             sequences[i][min(t + 1, seq_length - 1)] = eos_token_id
         #             break
         #
-        eos_indices = seq_length - attention_mask.fliplr().argmax(dim=1, keepdim=True).clamp(min=1)
+        eos_indices = seq_length - attention_mask.int().fliplr().argmax(dim=1, keepdim=True).clamp(min=1)
         attention_mask.scatter_(dim=1, index=eos_indices, value=1)
         sequences.scatter_(dim=1, index=eos_indices, value=eos_token_id)
